Ive done my part - Ian
Implemented Database class
Wrote examples and documentation on how to use each class that ive written

+++ src/javameowmeow/DataRow.java
@@ -1,7 +1,7 @@
 package javameowmeow;
 
 public class DataRow {
-	int ID; //used to get id of row ID of database
+	long ID; //used to get id of row ID of database
 	String source; //eg. facebook, reddit, twitter
 	String article_id; //string to identify the post
 	String article; //eg. news article, post
